requirements/base.txt
# Development requirements

Django==1.7.4
django-polymorphic==0.6.1

django-xworkflows==0.9.4
xworkflows==1.0.4
django-uuidfield==0.5.0
